refactor: improve error handling (AM-215) (#101)
