+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/AccessManagementService.java
@@ -64,27 +64,24 @@ public class AccessManagementService {
      * <p>Operation is performed in a transaction so that if not all records can be created then whole grant will fail.
      *
      * @param explicitAccessGrant an object that describes explicit access to resource
+     * @throws PersistenceException if any persistence errors were encountered causing transaction rollback
      */
     public void grantExplicitResourceAccess(@NotNull @Valid ExplicitAccessGrant explicitAccessGrant) {
         jdbi.useTransaction(handle -> {
             AccessManagementRepository dao = handle.attach(AccessManagementRepository.class);
-            try {
-                explicitAccessGrant.getAttributePermissions().entrySet().stream().map(attributePermission ->
-                    ExplicitAccessRecord.builder()
-                        .resourceId(explicitAccessGrant.getResourceId())
-                        .accessorId(explicitAccessGrant.getAccessorId())
-                        .permissions(attributePermission.getValue())
-                        .accessType(explicitAccessGrant.getAccessType())
-                        .serviceName(explicitAccessGrant.getServiceName())
-                        .resourceType(explicitAccessGrant.getResourceType())
-                        .resourceName(explicitAccessGrant.getResourceName())
-                        .attribute(attributePermission.getKey())
-                        .securityClassification(explicitAccessGrant.getSecurityClassification())
-                        .build())
-                    .forEach(dao::createAccessManagementRecord);
-            } catch (Exception e) {
-                throw new PersistenceException(e);
-            }
+            explicitAccessGrant.getAttributePermissions().entrySet().stream().map(attributePermission ->
+                ExplicitAccessRecord.builder()
+                    .resourceId(explicitAccessGrant.getResourceId())
+                    .accessorId(explicitAccessGrant.getAccessorId())
+                    .permissions(attributePermission.getValue())
+                    .accessType(explicitAccessGrant.getAccessType())
+                    .serviceName(explicitAccessGrant.getServiceName())
+                    .resourceType(explicitAccessGrant.getResourceType())
+                    .resourceName(explicitAccessGrant.getResourceName())
+                    .attribute(attributePermission.getKey())
+                    .securityClassification(explicitAccessGrant.getSecurityClassification())
+                    .build())
+                .forEach(dao::createAccessManagementRecord);
         });
     }
 
@@ -94,7 +91,8 @@ public class AccessManagementService {
      * <p>IMPORTANT: This is a cascade delete function and so if called on a specific attribute
      * it will remove specified attribute and all children attributes.
      *
-     * @param explicitAccessMetadata an object to remove a specific explicit access record.
+     * @param explicitAccessMetadata an object to remove a specific explicit access record
+     * @throws PersistenceException if any persistence errors were encountered
      */
     public void revokeResourceAccess(@NotNull @Valid ExplicitAccessMetadata explicitAccessMetadata) {
         jdbi.useExtension(AccessManagementRepository.class,
@@ -106,7 +104,8 @@ public class AccessManagementService {
      *
      * @param userId     (accessorId)
      * @param resourceId resource Id
-     * @return List of user ids (accessor id) or null
+     * @return list of user ids (accessor id) or null
+     * @throws PersistenceException if any persistence errors were encountered
      */
     public List<String> getAccessorsList(String userId, String resourceId) {
         return jdbi.withExtension(AccessManagementRepository.class, dao -> {
@@ -124,7 +123,8 @@ public class AccessManagementService {
      * @param userRoles accessor roles
      * @param resources envelope {@link Resource} and corresponding metadata
      * @return envelope list of {@link FilterResourceResponse} with resource ID, filtered JSON and map of permissions
-     *     if access to resource is configured, otherwise null.
+     *     if access to resource is configured, otherwise null
+     * @throws PersistenceException if any persistence errors were encountered
      */
     public List<FilterResourceResponse> filterResource(@NotBlank String userId,
                                                        @NotEmpty Set<@NotBlank String> userRoles,
@@ -142,7 +142,8 @@ public class AccessManagementService {
      * @param userRoles accessor roles
      * @param resource  envelope {@link Resource} and corresponding metadata
      * @return envelope {@link FilterResourceResponse} with resource ID, filtered JSON and map of permissions if access
-     *     to resource is configured, otherwise null.
+     *     to resource is configured, otherwise null
+     * @throws PersistenceException if any persistence errors were encountered
      */
     @SuppressWarnings("PMD") // AvoidLiteralsInIfCondition: magic number used until multiple roles are supported
     public FilterResourceResponse filterResource(@NotBlank String userId,
@@ -203,6 +204,7 @@ public class AccessManagementService {
      * @param roleNames    A set of role names. Currently only one role name is supported but
      *                     in future implementations we shall support having multiple role names
      * @return a map of attributes and their corresponding permissions or null
+     * @throws PersistenceException if any persistence errors were encountered
      */
     @SuppressWarnings("PMD") // AvoidLiteralsInIfCondition: magic number used until multiple roles are supported
     public Map<JsonPointer, Set<Permission>> getRolePermissions(@NotBlank String serviceName,
