+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/exceptions/UnsupportedPermissionsException.java
@@ -1,6 +1,9 @@
 package uk.gov.hmcts.reform.amlib.exceptions;
 
-public class UnsupportedPermissionsException extends RuntimeException {
+/**
+ * Thrown when access management was misconfigured with unknown permissions.
+ */
+public class UnsupportedPermissionsException extends AccessManagementException {
     private static final long serialVersionUID = 1L;
 
     public UnsupportedPermissionsException() {
