+++ am-lib/src/main/java/uk/gov/hmcts/reform/amlib/internal/aspects/AspectsSuite.java
@@ -1,0 +1,9 @@
+package uk.gov.hmcts.reform.amlib.internal.aspects;
+
+import org.aspectj.lang.annotation.Aspect;
+import org.aspectj.lang.annotation.DeclarePrecedence;
+
+@Aspect
+@DeclarePrecedence("ValidationAspect, ErrorHandlingAspect")
+public class AspectsSuite {
+}
